SolarNasih_SMA/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
httpx==0.25.0

# Modèles et validation
pydantic==2.5.0
pydantic-settings==2.1.0

# APIs externes (directement)
google-generativeai==0.3.2
tavily-python==0.3.0

# Utilitaires
python-dotenv==1.0.0
jinja2==3.1.2
requests==2.31.0
PyYAML==6.0.1

# Audio (optionnel)
speechrecognition==3.10.0
gtts==2.4.0

# Documents
python-docx==1.1.0
openpyxl==3.1.2
fpdf2==2.7.6

# Utilitaires finaux
markdown==3.5.1
beautifulsoup4==4.12.2

langgraph==0.0.26
langchain==0.1.0